Refactor attachment image preview section
